✨ feat(toolregistry): refactor OpenAPI integration to modular structure
- move OpenAPI client configuration to separate ClientConfig class
- create dedicated tool wrapper class for OpenAPI operations
- implement async/sync execution capabilities
- extract URL detection and schema loading logic to utils module
- reorganize integration logic into proper class structure
- update tool registration methods to use new modular components
